Computer Vision Dataset
Popular Computer Vision Datasets: ImageNet, Coco, and Google Open Images
Whenever you work on supervised machine learning problem, the biggest challenge is the collection of data. 
If we work on cat and dog image classification, we have to gather so many dog and cat images. We have to annotate them, put label on them, this is the cat and that is the dog.
The deep learning need a huge amount of dataset. If the dataset is bigger, then generally the deep learning is effective.