Vanish and Explode Gradient
1. What is difference of Gradient and Derivative?
1.1 Gradient is vector and Derivative is scalar.
1.2 Gradient can be the overall multiplication of chain rule of individual derivatives.
2. What is vanish gradient
Since the gradient is the multiplcation of chain rule of each individual derivatives, 
one of the derivative is zero. then overall giadeient will be zero.
When we update the new_weight = old_weight - learning_reate * gradient,
it will be very slow. Almost learning nothong. Just lise a dumb student in the class without learn anything.  
We call this concept is "Vanishing Gradient".
3. On the other hand, if some of the derivative is big, the overall multiplication will become huge.
We call this concept is "Exploding Gradient".
4. In RNN, the input statement change can be very small from sequence to sequence, 
we encounter "Vanishing Gradeint" problem all the time.
5. In RNN, we have GRU (Gated Recurrent Unit) and LSTM (Long Short Term Memory) to solve "Vanishing Gradeint" problem.