Full Stack Model
Full Stack Build Model:
We have load the dataset into the RAM Memory. 
1. The tensorflow cache() manage CPU hardware to open the file once instead of multiple times.
2. Th tensorflow prefetch() allowed us pipeline and parallel run between the CPU and GPU accelerators. 
We will build the CNN (Convolutional Neural Network) and train the model with our dataset.
We will measure the accuracy of model prediction with validate and test dataset. 
If the accuracy of model is good, we will export the model into the disk.
We can use the model with FastAPI Server.